Fix potential "clobbers" in longjmp
Any local variables used after a longjmp are potential victims of
"clobbering", because they may have been stored in registers.  One
can address this by making those local variables "static" or
"volatile", but this may not always have the semantics you want
(static variables can be bad for threads, volatile variables may
not let you pass the references to routines that are non-volatile).

In the case of the Init_Core()'s boot function, volatile is not an
option for the `boot` variable.  So the easiest way to solve this
problem is to move the code that does the setjmp/longjmp into its own
function where the offending variables are turned into parameters
that can't be clobbered:

http://stackoverflow.com/a/2105840/211160