Some naming and casing changes to think about
Some naming issues from R3-Alpha have been gradually phased out over
time.  For instance, to initialize a value cell with an integer there
was a macro SET_INTEGER(...).  But this gets quite confusing if you
try to expand that pattern to initialize a cell for a WORD!, because
you get SET_WORD(...).  It's an obviously broken pattern that has
given way to things like `Init_Block(...), `Init_Set_Word(...)`,
`Init_Word(...)`, etc.  So at some point the writing is on the wall
that SET_INTEGER will join the club and become `Init_Integer(...)`

But there are other less obvious cases, and this is an incomplete bit
of change to spark the thought experiment of considering them.

R3-Alpha had quite a lot of preprocessor macros doing its work.  This
was problematic for all the reasons that preprocessor macros are
usually problematic...they don't have any type signatures, if you
pass them something with side effects (like a function call) and they
use that parameter more than once, you'll get the side effects multiple
times, etc. etc.

Ren-C bit the bullet and was completely re-laid out in a way that had
the dependencies in line to allow for inline functions.  In most all
circumstances, nothing is allowed to be a macro if it would repeat
its arguments...if macro features are absolutely necessary (e.g.
capturing __FILE__ and __LINE__) then it is broken into two parts--a
macro and an inline function that macro calls.

Yet the C convention of warning people something is a macro by using
ALL-CAPS was not abandoned at the moment of that transition.  This
leaves the question on the table of exactly how should the "tool" of
naming, capitalization, casing, etc. be used in the codebase.

It's hard to tell exactly what's good or bad if one is used to looking
at things a certain way...sometimes it takes time.  But if you are
reading something like:

    if (NOT_END(f->value)) { ... }

and it gets changed to:

    if (Not_End(f->value)) { ... }

It might feel like "tests are all caps, variables are lowercase"
had provided an element of contrast and readability.  Regardless of
if NOT_END is a function or a macro, it might be more "operator-like"
and the capitalization might be a good thing.

I'm not sure on the answers, so this is really just a conversation
piece for the moment.  I know I like Init_Word() better than SET_WORD()
and I'm fairly sure I like Init_Word() better than INIT_WORD(), but I
don't know if I like Is_Word() better than IS_WORD() yet.