Beginnings of userspace encap, ELF format add/update
This is the start of work on improving encapping, by removing it from
the C code and putting it into the host, such that the encap tool and
the encap checks are done by platform-independent Rebol code.  Hence
a Windows Rebol executable could encap a Linux Rebol executable that
was acquired from somewhere else.

Though this commit only contains code for ELF, @ShixinZeng is working
on a Windows PE version.