Unify semantics of GET-PATH + PICK*, and SET-PATH! + POKE
There was significant common code between PICK and POKE and the
GET-PATH! and SET-PATH! operations, with minor differences between
the operations that weren't clearly intentional.

It is not possible to write path processing in terms of PICK and POKE
such that `a/b/c: d` would act as `poke (pick (pick a 'b) 'c) d`.
The reason for this is that the result of the picks is a value taken
out of a position, which no longer has a relationship to the location
from which it was taken.

However, it is mostly possible to write PICK and POKE in terms of path
processing.  The exception is if PICK and POKE are operations used
as port actions, because there is no way for ports to override path
dispatch...and if they did, they would lose the ability to behave as
objects.  (Note: this points to a bigger issue about ports not being
able to serve two masters, as many actions that apply to objects could
be interesting when defining specialized port actions, e.g. APPEND)

This commit forces PICK and POKE through the same code path used by
path getting and setting.  That code is known to be needing attention
to improve it, but this helps ensure that the improvement can benefit
(and be checked) by both operating forms.

In order to work around the current PORT! issue, the SYM_PICK_P and
SYM_POKE symbols are sent explicitly to the port actor.

The most significant impact of this commit that people should notice
is that `blk/picker` will error if the picker does not pick out of
a block, just as `obj/picker` would do if there was nothing picked
out of an object (or `var` would do if there was no variable).  One has
to use `:blk/picker`.  Examining the affected code shows that it is
an apparent benefit to need to distinguish whether you expect a value
or not, but that the hard-to-handle nature of void means one often
has to say `to-value :blk/picker` or other mitigation to get a blank.
This suggests a need to investigate alternatives, e.g. if @blk/picker
could be a request for a void on absence, with :blk/picker being a
request for a blank.