BLANK! vs. error for `read dns://` when network works, but not found
This changes the READ DNS:// behavior to give errors when a truly
exceptional network problem happens, but to give blanks when a lookup
connects to the name server but just fails.

It also gets rid of the asynchronous DNS lookup code on Windows.  The
relevant API calls were not carried forward with IPv6, and are thus
deprecated.  Microsoft "urges" that the more standard/POSIX network API
be used with a thread if this capability is desired.