TO-LOGIC -> native, kill TRUE?/FALSE?, add TRUTHY?/FALSEY?
TRUE? and FALSE? are too misleading to exist.  This changes the core to
use TO-LOGIC and NOT, which is probably what everyone should do.  But
to some people's tastes, they may prefer TRUTHY? and FALSEY? to convey
the meaning.  Add those and let people experiment with them to see if
they really can stand them and like them better than TO-LOGIC or NOT.