Splice only BLOCK! by default when no /ONLY (APPEND, INSERT)
This is a proposed simplification which would make the splicing
behavior of APPEND, INSERT, and their dependent operations only apply
to BLOCK!.  Hence ANY-PATH! and GROUP! would not splice.

    >> data: copy [a b c]

    >> append data 'd/e
    == [a b c d/e]

    >> append data quote (f g)
    == [a b c d/e (f g)]

    >> append data [h i]
    == [a b c d/e (f g) h i]

    >> append/only data [j k]
    == [a b c d/e (f g) h i [j k]]

Surprisingly (or perhaps unsurprisingly) this didn't seem to have any
obvious effect on the code.  It is very rarely desirable to splice a
PATH! or GROUP! into a block, and most operations that wish to add new
elements onto a path or group do so either with items one-at-a-time or
with a BLOCK! to splice.

One aspect of this change is that it makes it easier to produce paths
nested inside of paths.  The system does not handle this very well in
rendering, but it needs to, as they've always been quite possible to
create.  Pursuing an answer to that is discussed here:

https://forum.rebol.info/t/226/3