Discern anonymous functions from ones named --anonymous--
Historically, there has been a symbol --anonymous-- which was used in
slots where a function name might appear but there isn't a relevant
one. For instance:

    takes-int: func [x [integer!]] [...]
    do compose [(:takes-int) #not-an-integer]

The error message would say "--anonymous-- does not allow TAG! for its
x argument", with the WORD! --anonymous-- fabricated as the parameter
to the error.

Under this scheme, there's no way when interpreting the error/etc.
as talking about a function whose name is *actually* --anonymous-- vs.
one that has been made up.  It also feels rather sketchy to be writing
`label = '--anonymous--` as that test ("wait, how many dashes was it?")

This commit switches to actually having the stack frames distinguish
between having a symbol and not, and when reflecting the anonymous
frames will come back with a BLANK!.  In order to avoid too many sites
having to process the potential NULL of the frame label, operations to
derive UTF-8 string names or a WORD!/BLANK! are generalized vs. the
direct extraction of the label field...leading some routines that used
to take labels to take frames instead.

Also since function frames have been unified, this renames the
"Is_Any_Function_Frame()" test to simply "Is_Function_Frame()"