Make EITHER always use "EITHER/ONLY" semantics
Unlike conditional constructs that may decide not to run a branch,
either *always* runs a branch.  Using it with ELSE or THEN is thus
pointless, so interfering with the branch return results through the
"blankification" interferes with `result: either x [void] [something]`
for no particular benefit.

This proposal changes things so that rather than cripple EITHER
just for the sake of making it 100% interchangeable with
`result: if x [void] else [something]`, it is allowed to be different.