Experimental implementation of `<skip>` function arguments
This is a very simplistic hack to try and implement skippable function
parameters, to see how useful or interesting they are...before trying
to go through a code reorganization to support them more legitimately.

Arguments marked `<skip>` will--instead of generating a type error
when a parameter cannot match--set that argument to void and let the
parameter "roll over" to the next compatible parameter.  (e.g. a
quote can roll over to a quote, but an evaluated parameter cannot
roll over to a quote because it's too late.)

    emit: function [file! [<skip> file!] block [block!]] [...]

    emit [x y z]
    emit %foo.txt [x y z]

While this could already be done with variadics, including this more
narrow interface makes it possible to use the function normally with
APPLY, and it is easier to write.

Because this case is only handled when an argument has a type mismatch,
the way the feature is implemented here does not affect the speed of
the evaluator.  More invasive changes would be required if someone
comes up with motivating examples for this feature.  It could be
possible to roll over quoted parameters to then perform an evaluation
using the quoted data