"Latin-1 Nowhere", use REBVAL* to exchange FILE! with host
This is a *significant* change which is difficult to break into smaller
pieces...because it removes the existence of byte-sized ANY-STRING!
entirely.  It means basically anything that had conditional behavior
on VAL_BYTE_SIZE(str) had to be changed.

Consolidating on wide strings is preparation for UTF8-Everywhere, and
it made sense to go with the wide string case branch because that has
to handle large codepoints...as all UTF-8 strings are going to need to
be able to do.  Unsurprisingly, it found bugs in some routines you
would have only noticed if you happened to use them with wide strings.

While that might sound like an unremarkable change on the surface
(deleting the Latin1-byte-sized-string code paths), it's rather
involved due to building a firmer distinction between BINARY! and
STRING!, as though they both will be "byte-sized" after the UTF-8
Everywhere conversion, they will behave very differently.

One sticking point was REBCHR and its role in using byte-sized string
values to back host interactions on POSIX.  Fixing that meant pushing
through the idea that the host uses REBVAL* as currency.  So interfaces
that used to expect either a char* or a wchar_t* based on whether you
were on windows or not, now take a REBVAL* and get FILE!s or STRING!s,
and then use libRebol APIs to extract the kind of data they want.

On the very plus side of that, REBCHR is basically eliminated.