Use UTF-8 for the mold buffer
The easiest target for classically REBUNI-based code to convert to
UTF-8 is the mold buffer.  This is because it is only appended to with
material popped off.  Since nothing is inserted, the irregular length
of encoded codepoints isn't a big issue.  Yet molding/forming comprises
a very large portion of Rebol's code detailed with the implementation
concerns of strings.

There was some amount of detailed calculation about buffer size
expansions for the mold buffer that this commit replaces with guesses,
and those points are noted with Prep_Mold_Overestimated.  The guesses
are derived from the prior guesses, but rounding up to a worst case
scenario of every codepoint requiring 4 bytes in UTF-8.  (Expanding the
mold buffer conservatively isn't really a significant problem, given
that the space is reused.  This can be revisited later.)