UCS2 absorbed into RL_API, FAIL hook, API retooling
In trying to advance the agenda of UTF8-Anywhere, one area that has
to be clamped down on is any code outside of the core which reaches
into the internals of how strings are stored internally.  The ODBC
extension tried to leverage the idea that there were "UCS-2" strings
supported in the core to be able to deal with a 16-bit-per-character
representation on both Windows, and Linux (where wchar_t is not
necessarily 16-bits).

Motivated by the desire to excise dependency of ODBC on Rebol
internals, this pushes and documents the gentrifying RL_API a bit
further...with naming changes, comment rewrites, and general cleanup.
The now-UTF8-based mold buffer is used as the lingua-franca of how
alternative representations are converted into "Rebol strings"--
however Rebol strings are implemented.

As a result, some RL_API UCS2 routines are introduced.  (It is likely
that wchar_t APIs are not relevant by comparison, and actual UTF-16
support is what's really needed.)  The ODBC code's dependence on
REBSERs or UNI_HEAD(...) on series is excised completely, along with
the introduction of several other basic routines to lessen the
dependency on the internal API.

Many unused APIs are deleted, including GOB-related ones (which now
the R3-View code would presumably be better served by using the
internal API directly.)

One interesting change is that rebFail() now delegates to a FAIL native
which is hijacked in the mezzanine by a userspace primitive.  This is a
pattern similar to what rebPanic() does.