Clipboard uses REBVAL* as currency, more UTF-8 mothballing
The deprecated REBREQ method of exchanging data with the host "Device"
for the clipboard used a roundabout method of passing data, involving
a raw byte pointer and an RRF_WIDE flag describing whether the buffer
carried wide character data.  Consistent with the emerging cleanup in
other code, this changes the protocol to just exchange a REBVAL* of
a binary built via the API, and then does conversions at the endpoint.

Includes deletions and reorganization of various bits of internal
unicode handling that were liberated by the change.