Fix rebFail() bug, put rebFail_OS() in API temporarily
The idea behind rebFail() was to be able to dispatch out through the
a native FAIL FUNCTION!, giving users the opportunity to hijack the
failure.  It would start as an experiment in the API and then perhaps
be integrated into fail() more generally.

Interesting idea though it may have been, it didn't have much in the
way of usages, and was never tested.  Since it relied on rebDo() to
get its work done, and it called FAIL, that rebDo() would always
trap it and return NULL.  So the core fail would have to be called in
all cases, after that trap.

This fixes that problem.  It also makes a small interim concession to
add OS-specific code to the API to handle GetLastError() or errno
signals, just to provide a common fail hook for extensions that want
to give literate error messages when those things happen.  Ideally
this shareable routine could be shared another way without making the
core OS-dependent.