Use API handles in timestamp API, RXT_ERROR, Init_Handle
One holdover from when actual REBVAL cells were passed around to the
host API was the timestamp-oriented API.  This was an interim step
up from use of a structure called REBOL_DAT, but predated the curent
API notion that the currency of exchange with the "host" or other
extensions was "const REBVAL *" as opposed to cells to be written
into.  This changes the API to use Alloc_Value(), though more thought
should be put into the exact nuances of the time and date API.

Also adds RXT_ERROR as an external type, and makes Init_Handle_XXX
routines return the handle initialized--consistent with a prior
commit about making it easier to avoid intermediate variables.