Get rid of WM_DNS
R3-Alpha used WSAAsyncGetHostByName and WSAAsyncGetHostByName to do
non-blocking DNS lookup on Windows.  These functions are deprecated,
since they do not have IPv6 equivalents...so applications that want
asynchronous lookup are expected to use their own threads and call
getnameinfo().

It complicated the system for a Windows-only feature and had already
been removed, but the dependent code (Done_Device with a potential
error code) was still hanging around.  This eliminates it completely.