Replace DR_ERROR and DEVREQ.error integer code with rebFail()
The R3-Alpha model of operating system abstraction ("device requests") took in a DEVREQ C structure of C data types, and returned either DR_DONE (synchronously completed) DR_PENDING (in the event queue) or DR_ERROR.  The error state was communicated by an integer-based code filled into the DEVREQ.  This did not provide clear error messages, and many were made up numbers like -18.

This removes that mechanic and switches to use rebFail(), which is an API-exported version of the internal fail() routine.  Given the commonness of Windows and POSIX error code handling, it also puts a small temporary amount of OS-specific code into the API core, so that OS failures will read more meaningfully.

This additionally adds a rebManage() function to the API in order to say that an API value will be freed by its stack frame.  It is still possible to release a value even if it's managed.  The default behavior for the API in the future will be to manage handles unless asked otherwise...but just for the sake of bookkeeping they are currently manual by default and leak checked.