Merge UCS2 and "wide character" APIs, deprecating wchar_t
The wchar_t type is not a standard size on non-Windows platforms, and most Linux and POSIX code uses UTF-8 (or if older, can get by with ASCII).  Since wchar_t predates unicode, its usefulness is not that high...the only relevant "wide character" code on Linux seems to be that which is trying to be Windows-compatible, using a short int definition (like SQLWCHAR).

This suggests merging the UCS2 routines and "wide character" routines to be the only non-UTF8-supported APIs.   However, Windows defines its WCHAR type as wchar_t.  So if one is building with C++, it is not legal to simply take a 16-bit unsigned of any other type and make that the interface and allow callers to use it without a cast.

This makes a compromise by having a platform-dependent definition for REBWCHAR to use in the interface, that is wchar_t on Windows and a 16-bit integer on other platforms.  This isn't intended to be a type anyone uses in their code--if one is writing on Windows they should use WCHAR and on Linux, whatever type they are trying to interface with (e.g. SQLWCHAR).

This also eliminates all wchar_t references in the Windows code, replacing them with WCHAR--to help discourage the use of the datatype in the system.

Also trims down some wide string specific code and removes it from the system.