Move shared pointer detecting code for va_list into subroutine
The C language does not allow the first parameter of a function to be variadic.  This means one either must put a dummy parameter at the beginning, or have an abstraction which virtually "splices" the first element into a stream with the va_list.  

The Ren-C evaluator has a basis for this concept of a "prefetch" that allows one element to be virtually present at the beginning of an evaluation.  However, the existing code assumed that this element was a plain REBVAL...not something like a UTF-8 string.  This means the portion of the code that was designed to do the preload had to run through the same machinery that deals with fetching in the va_list normally.

Not being able to do this previously meant there were some spurious `rebDo(BLANK_VALUE, ...);` calls...which just had a throwaway blank value in the first slot.  This means they could be removed.

Also adds a rebAppend API, just so that the process module wouldn't be working with string internals.