Split Modify_String and Modify_Binary, delete Insert_String
The core modify routines behind INSERT, APPEND, and CHANGE on string and binary values shared a fair amount of code, though there were variances in the semantics of how the data would be interpreted.  However, with UTF-8 Everywhere, string and binary handling are drifting apart due to the lack of uniform codepoint size in strings.

This splits the routine apart, and also works on making the interface speak in terms of a REBVAL for the destination instead of a REBSER and an index, paving the way to utilize the caching mechanism.

The change brought to lite some breakage of using a UTF-8 mold buffer, when that buffer is also used for BINARY! set operations...which is now more apparent.  The likely answer for that and other binary operations will be to not use the mold buffer.