Move ENTAB and DETAB into native
The factoring out of entab and detab's functionality into C functions that operate on series instead of values interferes with the ability to optimize them based on cached UTF-8 offsets in the values.  Also, there were no other places that called the routines (it would not be a terrible amount of overhead to invoke it through the evaluator.

This also adds some new preparatory APIs for actually dealing with the "size" vs "length" distinction...and clarifies VAL_BIN calls by making them VAL_BIN_HEAD().