Change Copy_String_At_Len interface to take a REBVAL
The code for copying strings up to a limit will eventually benefit from the ability to have a REBVAL* with cached information about the offset locations of non-head codepoints.  This passes in the source string as a value instead of just a series (although some callers currently have to synthesize that value).

Also changes a trim function used only by Mold to operate on a byte-by-byte basis.