Get rid of direct canon pointer field in REBSTR
To accelerate the speed at which non-canon forms of strings could get
to their canon forms, one of the two "extra" pointers in the REBSER
node for a UTF-8 string (as used to back a WORD!) was used to point
directly at the canon form.

However, the other pointer was to a circularly linked list of all the
spelling forms, which could be enumerated to find a canon.  Hence this
was a relatively obscure optimization.

* Whether there is a canon flag or not on the form you are looking up
  had to be tested anyway at least once to know if you could use the
  optimized canon.

* Most of the time, you'll be hitting a canon form first, as people
  do most words in lowercase or whatever the main form is.

So this really only offered benefit in the case where you have more
than two spellings of the same word, and unless you have some large
number of variant casings of the spelling (dozens?) the amount of
benefit was extremely negligbile.

This is a step designed toward reclaiming the MISC() field to store
the length in *codepoints*.