Add AHEAD keyword, fix crash when using INTO during string parse
This adds the AHEAD keyword to PARSE.  Its lookahead behavior means
it's a more intuitive synonym for the AND operation.

It also fixes a crash that was occuring if you were already in the
middle of a string parse, and tried to invoke an INTO operation.  The
correct behavior in that case is to raise an error, as you cannot
currently parse "into" a character.

Tidies up a couple of redundant initializations in the simulated
function application used for SUBPARSE-based frames...which should be
revisited in light of the desire to be able to HIJACK parse recursions
for debugging (as in the ENCLOSE/HIJACK example)