Make Is_Value_Managed() tolerate unreadable blanks
When the evaluator no longer can use an output cell for meaningful
input to the next evaluative step (e.g. it is not intentionally input
to the left side of an enfix operation) the old not-to-be-used content
is masked as a BLANK! with the "unreadable" bit set on it.  A trash
cell (uninitialized) cannot be used because the cell is expected to
be protected by the GC.  And an end cell cannot be used because that
denotes the special state of "no data" which is needed for proper
processing of invisibles/elide.

So an unreadable blank is used instead.  It is a debug-build-only bit
set on BLANK! which makes it GC safe, but attempts to read it (even
asking for its type) will cause an alarm.  This bit is explicitly used
to create interim values that are GC safe (as opposed to trash), but
which are not actually representative of finished data that should be
read from.

The Is_Value_Managed() routine is largely intended for use by the GC,
many circumstances of which are explicitly supposed to tolerate the
"unreadable" BLANK! bit.  This makes it tolerant of unreadable blanks,
and also goes ahead and shortens up the code using shorter variable
names for the only relevant values in the routine.