Non-bounded SKIP/ONLY, NEXT+BACK don't clip (yield blank)
SKIP and AT have historically clipped their skipping at the head and
tail of the series, e.g.

    >> skip "abcd" -100
    == "abcd"

While arguably useful, it's not always the intent.  This adds an /ONLY
refinement that does not bound in this way, but returns blank if it
tries to go past the bounds:

    >> skip/only "abcd" -100
    == _

The NEXT and BACK specializations (skip offset +1 and -1 respectively)
are then modified to use this invariant, as it is less likely to
obscure bugs...and can make tests for TAIL? or HEAD? unnecessary:

    while [pos: next pos] [
        ...
    ]

If one truly needs the previous bounding behavior, SKIP without /ONLY
can be used with 1 or -1 respectively.

This also revises the treatment of LOGIC! with SKIP, such that it
presumes TRUE means act like NEXT, while FALSE means return the series
at its current position.

    >> skip "abc" true
    == "bc"

    >> skip "abc" false
    == "abc"