BLANK! triggers ELSE (vs. just void), EITHER acts as in R3-Alpha
ELSE was designed to examine the evaluated value on its left, and pass
it through if it was non-void...otherwise evaluate the branch it was
provided.  When combined with a convention of conditional constructs
only returning void if they didn't run any branches, this could give
the "illusion" of a conventional language's behavior running ELSE
clauses.  (The illusion was enhanced with other enfix operators such
as ALSO and THEN.)

But since BLANK! was considered a value to be passed through vs. one
that the ELSE would branch on, it meant it could not be used with
constructs that returned BLANK! to indicate a "miss".  There are lots
of those now, like ALL, ANY, SELECT, PICK, FIND, etc...and users would
likely be using the convention in the future to make many more.

This commit tweaks the balance so that ELSE will trigger off of a
blank, as well as a void.  Then when a conditional takes a branch, if
that branch evaluates to *either a void or a blank*, it will modify the
result to be a BAR!.  This means:

    >> if true [1020]
    == 1020

    >> if true [print "Hi"]
    == |

    >> if true [_]
    == |

BAR! is chosen specifically for its "jarring" status as a value that
is uncommon, and doesn't gloss over the potential misunderstanding
(the way a #[false] might, since it is falsey *like* a blank while
clearly not being one).  It also is light to look at in the console
as an expression result.

The casualty of this decision is that constructs like CASE or SWITCH
cannot deliberately return BLANK! values unless variants e.g. `CASE*`
or `SWITCH*` are used (which instruct them to not alter the branch
results).  The option already existed for the purposes of passing
through voids unchanged, but this adds blanks to it.

It's a change, but not as scary as it first appears.  Other parts of
Rebol, such as APPEND, ask you to tack on an /ONLY if you want to mean
"verbatim".  And so this simply means that if you're writing a
conditional expression with branches that are returning a NOTHING?
(e.g. a void or a BLANK!), and that's what you meant, you use a
refinement or a one-character shorthand to say so.

There are mitigating factors for this.  For instance, one reason that
a conditional might like to return a blank deliberately would be to
run some code but try and signal to remove its effect, e.g.

    print [
       "stuff"
       if condition [
           do-some-diagnostics
           drink-some-tea
           blank
       ]
       "more stuff"
    ]

If the desire is to never inject content, that can be achieved now
with `elide (if condition [...])`.

Those wishing to deliberately return blanks in a conditional sense
might find rewriting their IF clauses as ALL clauses works for them,
as a "miss" on ALL will return a blank.

As an additional mitigation, EITHER is removed from acting like this
group of functions...and returns the result of its branch evaluations
literally.  This gives an anchor for many common cases, in such a way
that code can work across Rebols.  It does so at the cost of not making
it always legal to change EITHER into IF ... ELSE.