Add runtime stack growth direction detection option
R3-Alpha's method of doing stack overflow detection was based on a
non-standard C technique...which wasn't really its fault, because there
is no mention of a "stack" at all in the C standard.

In order to try and preempt common crashes, recursive routines would
conservatively test pointers of their local variables against a
"stack limit" calculated against a boundary location in memory.  While
some C compilers grow their stacks up and others down, down was most
common.  While this is not a perfect method, it's one of the few
available options in C programs that want to try not to crash in the
face of a C stack overflow.

R3-Alpha tried to detect the stack growth direction using a comparison
of addresses of local variables, which broke in some optimizers.  It
was removed due to the uncommonness of stacks growing up, in favor of
putting the direction in the build configuration.

However, Emscripten and webassembly don't have a firm commitment on
which direction the stack grows:

https://github.com/kripken/emscripten/issues/5410

So for now, this creates an option of a semi-reliable detection via
a routine that is recursive in an optimizer-unfriendly way.  A build
configuration can specify this detection, or that the stack grows
up or down.